- Implemented book creation feature - Added docker configuration for angular application - Adjusted SecurityConfig files on auth and resource server to allow access to SPA client

+++ book-library-frontend/src/app/components/book/book.component.ts
@@ -1,15 +1,16 @@
 import {Component, OnInit} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {FormsModule} from '@angular/forms';
+import {FormsModule, FormGroup, FormBuilder, Validators, ReactiveFormsModule} from '@angular/forms';
 import {Router} from '@angular/router';
 import {AuthService} from '../../services/auth.service';
 import {BooksService} from '../../services/book.service';
 import {Book} from '../../model/book.model';
+import {BookRequest} from '../../model/book.request.model';
 
 @Component({
   selector: 'app-books',
   standalone: true,
-  imports: [CommonModule, FormsModule],
+  imports: [CommonModule, FormsModule, ReactiveFormsModule],
   templateUrl: './book.component.html',
   styleUrl: './book.component.css'
 })
@@ -23,12 +24,22 @@ export class BookComponent implements OnInit {
   singleBookResult: Book | null = null;
   categories: string[] = [];
   authors: string[] = [];
+  bookForm: FormGroup;
+  showBookForm: boolean = false;
 
   constructor(
     private readonly booksService: BooksService,
     private readonly authService: AuthService,
-    private readonly router: Router
+    private readonly router: Router,
+    private readonly formBuilder: FormBuilder
   ) {
+    this.bookForm = this.formBuilder.group({
+      title: ['', [Validators.required, Validators.minLength(3)]],
+      author: ['', [Validators.required, Validators.minLength(3)]],
+      amountPages: [0, [Validators.required, Validators.min(1)]],
+      category: ['', Validators.required],
+      content: ['', [Validators.required, Validators.minLength(10)]]
+    });
     this.loadCategories();
     this.loadAuthors();
   }
@@ -106,9 +117,26 @@ export class BookComponent implements OnInit {
     this.selectedBook = null;
   }
 
-  createBook(book: Book) {
-    this.booksService.createBook(book).subscribe({
-      next: () => {
+  createBook(): void {
+    if (this.bookForm.invalid) {
+      Object.keys(this.bookForm.controls).forEach(key => {
+        const control = this.bookForm.get(key);
+        control?.markAsTouched();
+      });
+      return;
+    }
+
+    const newBook: BookRequest = {
+      title: this.bookForm.get('title')?.value,
+      author: this.bookForm.get('author')?.value,
+      amountPages: this.bookForm.get('amountPages')?.value,
+      category: this.bookForm.get('category')?.value,
+      content: this.bookForm.get('content')?.value
+    };
+
+    this.booksService.createBook(newBook).subscribe({
+      next: (createdBook) => {
+        this.toggleBookForm();
         this.searchBooks();
       },
       error: (err) => this.handleError(err)
@@ -116,15 +144,17 @@ export class BookComponent implements OnInit {
   }
 
   deleteBook(id: number) {
-    this.booksService.deleteBook(id).subscribe({
-      next: () => {
-        this.searchBooks();
-        if (this.selectedBook?.id === id) {
-          this.selectedBook = null;
-        }
-      },
-      error: (err: any) => this.handleError(err)
-    });
+    if (confirm('¿Estás seguro de que quieres eliminar este libro?')) {
+      this.booksService.deleteBook(id).subscribe({
+        next: () => {
+          this.searchBooks();
+          if (this.selectedBook?.id === id) {
+            this.selectedBook = null;
+          }
+        },
+        error: (err: any) => this.handleError(err)
+      });
+    }
   }
 
   handleError(err: any) {
@@ -139,5 +169,12 @@ export class BookComponent implements OnInit {
 
   logout() {
     this.authService.logout();
+  }
+
+  toggleBookForm(): void {
+    this.showBookForm = !this.showBookForm;
+    if (!this.showBookForm) {
+      this.bookForm.reset();
+    }
   }
 }
